A generic TrackMate global spot detector that uses Appose.
This is a global detector, supposed to be operating on all the
frames of the input movie. The python script is provided with the
input image, and a dictionary of axis indices ('T':1, etc).
The python script is supposed to return the results of the segmentation
in a 'masks' image passed in the 'tasks.outputs' map. These masks
are then processed by the LabelImageDetector to create spots.