Generalize Mamba env config to any env config
Appose tries to be smart and detect what kind of config file content
it is: environment.yml, pixi.toml, etc. So we should not need to
hardcode only support for environment.yml. And in fact we need to use a
pixi.toml for proper cross-platform hardware-accelerated StarDist 3D,
so that we get the correct CUDA library versions and such in all cases.